Extract TaskList Identity
