+++ src/language-server/task-list/identity/storage.ts
@@ -1,8 +1,8 @@
 import type { IdentityStorage } from '../../../langium-model-server/identity/storage'
 import { AbstractIdentityStorage } from '../../../langium-model-server/identity/storage'
 import type { TaskListDocument } from '../workspace/documents'
-import { Model } from './task-list-identity'
-import type { TaskListIdentityIndex } from './task-list-identity-index'
+import { Model } from './model'
+import type { TaskListIdentityIndex } from '.'
 import type * as source from '../lms/model'
 
 export class TaskListIdentityStorage extends AbstractIdentityStorage<source.Model, TaskListIdentityIndex, TaskListDocument> implements IdentityStorage {
