+++ src/language-server/task-list/lms/task-list-facade.ts
@@ -12,9 +12,9 @@ import * as id from '../../../langium-model-server/semantic/model'
 import type { LangiumModelServerServices } from '../../../langium-model-server/services'
 import type { LmsDocument } from '../../../langium-model-server/workspace/documents'
 import * as ast from '../../generated/ast'
-import * as identity from '../../task-list/semantic/task-list-identity'
+import * as identity from '../identity/model'
 import * as semantic from '../semantic/model'
-import type { TaskListIdentityIndex } from '../semantic/task-list-identity-index'
+import type { TaskListIdentityIndex } from '../identity'
 import type { TaskListDocument } from '../workspace/documents'
 import { isTaskListDocument } from '../workspace/documents'
 import { Model, Task, Transition } from './model'
