+++ src/language-server/task-list/task-list-module.ts
@@ -1,10 +1,10 @@
 import type { Module } from 'langium'
 import type { LangiumModelServerServices, PartialLangiumModelServerServices } from '../../langium-model-server/services'
 import { TaskListValidator } from '../task-list/validation/task-list-validation'
-import type { TaskListIdentityIndex } from './semantic/task-list-identity-index'
-import { TaskListIdentityManager } from './semantic/task-list-identity-manager'
+import type { TaskListIdentityIndex } from './identity'
+import { TaskListIdentityManager } from './identity/manager'
 import { TaskListIdentityReconciler } from './semantic/task-list-identity-reconciler'
-import { TaskListIdentityStorage } from './semantic/task-list-identity-storage'
+import { TaskListIdentityStorage } from './identity/storage'
 import { TaskListSemanticDomain } from './semantic/task-list-semantic-domain'
 import type * as source from './lms/model'
 import { TaskListLangiumModelServerFacade } from './lms/task-list-facade'
